feat: Add home route and handler
This commit adds a new route and handler for the home page. The `RegisterRoutes` function in the `home` package is called to register the route with the router. The handler function `handlerHome` renders the home view using the `views.Home()` function. If there is an error during rendering, it returns a validation error.